add Promotion API dynamic lockStock
