+++ src/main/java/com/skillup/domain/promotion/PromotionService.java
@@ -3,18 +3,24 @@ package com.skillup.domain.promotion;
 // this is where we connect to database, take userDomain in, and
 
 
+import com.skillup.domain.promotion.stockStrategy.StockOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.jooq.True;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.List;
 
+@SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection")
 @Service // create a Spring object to containers
 @Slf4j
 public class PromotionService {
     @Autowired
     PromotionRepository promotionRepository;
+
+    @Resource(name = "${promotion.stock-strategy}")
+    StockOperation stockOperation;
 
     public PromotionDomain createPromotion(PromotionDomain promotionDomain){
         promotionRepository.createPromotion(promotionDomain);
@@ -30,24 +36,6 @@ public class PromotionService {
     }
 
     public boolean lockStock(String id){
-        synchronized (this){
-            // 1 get available stock, check >= 1
-            PromotionDomain currentPromotionDomain = promotionRepository.getPromotionById(id);
-            if (currentPromotionDomain.getAvailableStock() <= 0){
-                return false;
-            }
-            log.info("----- Current available stock is {} -----", currentPromotionDomain.getAvailableStock());
-            // 2 available stock - 1, set lock stock + 1
-            currentPromotionDomain.setAvailableStock(currentPromotionDomain.getAvailableStock() - 1);
-            currentPromotionDomain.setLockStock(currentPromotionDomain.getLockStock() + 1);
-            log.info("----- Updated current available stock is {}, lock stock is {} -----"
-                    , currentPromotionDomain.getAvailableStock(), currentPromotionDomain.getLockStock());
-            // 3 save domain
-            promotionRepository.updatePromotion(currentPromotionDomain);
-            return true;
-
-        }
-
-
+        return stockOperation.lockStock(id);
     }
 }
