+++ src/main/java/com/skillup/domain/promotion/stockStrategy/StockOperation.java
@@ -1,0 +1,6 @@
+package com.skillup.domain.promotion.stockStrategy;
+
+public interface StockOperation {
+
+    boolean lockStock(String id);
+}
