+++ src/main/resources/application.yaml
@@ -8,9 +8,13 @@ spring:
     username: root
     password: rootroot
 
-  flyway:
-    baseline-on-migrate: false
-    enabled: true
-    ignore-migration-patterns:
-      - "*:missing"
-    outOfOrder: true
+flyway:
+  baseline-on-migrate: false
+  enabled: true
+  ignore-migration-patterns:
+    - "*:missing"
+  outOfOrder: true
+
+# customized config
+promotion:
+  stock-strategy: sync
